chore(test-suite): custom app.js separated test environment
