+++ be/server.js
@@ -2,18 +2,10 @@ const express = require('express');
 const cors = require('cors');
 const dotenv = require('dotenv');
 const connectDB = require('./src/config/db');
-const authController = require('./src/controller/authController');
-const storiesRoutes = require('./src/routes/storiesRoutes');
+const app = require('./src/app');
 
 dotenv.config();
 connectDB();
-
-const app = express();
-app.use(cors());
-app.use(express.json());
-
-app.use('/api/auth', authController);
-app.use('/api', storiesRoutes);
 
 const PORT = process.env.PORT || 5000;
 app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
